Refactoring current solution
Split findstatspage versions into separate modules
Split az queries into separate query file and separate service
